added scoring.
i think the scoring works as described in the assignment, but it seems a little unfair. when one card is already face up and i pick a card that has an emoji that is the same as a card i already saw, i lose a point. i can immediately gain back one point if i remember where its match was and pick them on my next turn, but this seems like an unfair penalty. the example in the assignment doesn't distinguish between a card being seen by index or by identifier, so my code assumes its by identifier. this could be easily changed if necessary.